chore(ai): 패키지 정리
