+++ src/main/java/com/example/konnect_backend/domain/ai/domain/entity/log/AnalysisHistory.java
@@ -1,4 +1,4 @@
-package com.example.konnect_backend.domain.ai.entity.log;
+package com.example.konnect_backend.domain.ai.domain.entity.log;
 
 import com.example.konnect_backend.domain.ai.type.FileType;
 import com.fasterxml.jackson.annotation.JsonFormat;
