+++ src/main/java/com/example/konnect_backend/domain/ai/domain/vo/ExtractedText.java
@@ -1,0 +1,4 @@
+package com.example.konnect_backend.domain.ai.domain.vo;
+
+public record ExtractedText(String text) {
+}
