+++ src/main/java/com/example/konnect_backend/domain/ai/dto/internal/GeminiCallResult.java
@@ -1,6 +1,6 @@
 package com.example.konnect_backend.domain.ai.dto.internal;
 
-import com.example.konnect_backend.domain.ai.model.vo.TokenUsage;
+import com.example.konnect_backend.domain.ai.domain.vo.TokenUsage;
 
 public record GeminiCallResult(String response, TokenUsage tokenUsage, long maxTokens, String model, String finishReason) {
 }
