+++ src/main/java/com/example/konnect_backend/domain/ai/dto/internal/PromptTemplateWithModelName.java
@@ -1,5 +1,5 @@
 package com.example.konnect_backend.domain.ai.dto.internal;
 
-import com.example.konnect_backend.domain.ai.entity.PromptTemplate;
+import com.example.konnect_backend.domain.ai.domain.entity.PromptTemplate;
 
 public record PromptTemplateWithModelName(PromptTemplate prompt, String modelName) {}
