+++ src/main/java/com/example/konnect_backend/domain/ai/dto/response/ModelResponse.java
@@ -1,6 +1,6 @@
 package com.example.konnect_backend.domain.ai.dto.response;
 
-import com.example.konnect_backend.domain.ai.entity.AiModel;
+import com.example.konnect_backend.domain.ai.domain.entity.AiModel;
 
 public record ModelResponse(Long id, String name, String useCase) {
     public static ModelResponse from(AiModel m) {
