--- src/main/java/com/example/konnect_backend/domain/ai/model/vo/ExtractedText.java
@@ -1,4 +1,0 @@
-package com.example.konnect_backend.domain.ai.model.vo;
-
-public record ExtractedText(String text) {
-}
