+++ src/main/java/com/example/konnect_backend/domain/ai/repository/LlmCallMetadataRepository.java
@@ -1,6 +1,6 @@
 package com.example.konnect_backend.domain.ai.repository;
 
-import com.example.konnect_backend.domain.ai.entity.log.LlmCallMetadata;
+import com.example.konnect_backend.domain.ai.domain.entity.log.LlmCallMetadata;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LlmCallMetadataRepository extends JpaRepository<LlmCallMetadata, Long> {
