+++ src/main/java/com/example/konnect_backend/domain/ai/repository/PromptTemplateRepository.java
@@ -2,10 +2,9 @@ package com.example.konnect_backend.domain.ai.repository;
 
 import com.example.konnect_backend.domain.ai.dto.internal.PromptSummary;
 import com.example.konnect_backend.domain.ai.dto.internal.PromptTemplateWithModelName;
-import com.example.konnect_backend.domain.ai.entity.PromptTemplate;
+import com.example.konnect_backend.domain.ai.domain.entity.PromptTemplate;
 import com.example.konnect_backend.domain.ai.type.PromptStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
