+++ src/main/java/com/example/konnect_backend/domain/ai/service/module/UnifiedExtractorModule.java
@@ -1,12 +1,12 @@
-package com.example.konnect_backend.domain.ai.service.pipeline.module;
+package com.example.konnect_backend.domain.ai.service.module;
 
 import com.example.konnect_backend.domain.ai.dto.internal.ExtractionResult;
 import com.example.konnect_backend.domain.ai.dto.internal.GeminiCallResult;
 import com.example.konnect_backend.domain.ai.dto.response.ExtractedScheduleDto;
-import com.example.konnect_backend.domain.ai.entity.PromptTemplate;
+import com.example.konnect_backend.domain.ai.domain.entity.PromptTemplate;
 import com.example.konnect_backend.domain.ai.infra.GeminiService;
-import com.example.konnect_backend.domain.ai.model.vo.TokenUsage;
-import com.example.konnect_backend.domain.ai.service.pipeline.PipelineContext;
+import com.example.konnect_backend.domain.ai.domain.vo.TokenUsage;
+import com.example.konnect_backend.domain.ai.domain.vo.PipelineContext;
 import com.example.konnect_backend.domain.ai.service.prompt.PromptTemplateResolver;
 import com.example.konnect_backend.domain.ai.util.PromptUtils;
 import com.fasterxml.jackson.core.type.TypeReference;
