+++ src/main/java/com/example/konnect_backend/domain/ai/service/pipeline/DocumentAnalysisPipeline.java
@@ -3,20 +3,20 @@ package com.example.konnect_backend.domain.ai.service.pipeline;
 import com.example.konnect_backend.domain.ai.dto.internal.ExtractionResult;
 import com.example.konnect_backend.domain.ai.dto.response.DifficultExpressionDto;
 import com.example.konnect_backend.domain.ai.dto.response.DocumentAnalysisResponse;
-import com.example.konnect_backend.domain.ai.entity.PromptTemplate;
-import com.example.konnect_backend.domain.ai.entity.log.AnalysisRequestLog;
-import com.example.konnect_backend.domain.ai.model.vo.ExtractedText;
-import com.example.konnect_backend.domain.ai.model.vo.TokenUsage;
-import com.example.konnect_backend.domain.ai.model.vo.UploadFile;
+import com.example.konnect_backend.domain.ai.domain.entity.PromptTemplate;
+import com.example.konnect_backend.domain.ai.domain.entity.log.AnalysisRequestLog;
+import com.example.konnect_backend.domain.ai.domain.vo.ExtractedText;
+import com.example.konnect_backend.domain.ai.domain.vo.PipelineContext;
+import com.example.konnect_backend.domain.ai.domain.vo.TokenUsage;
+import com.example.konnect_backend.domain.ai.domain.vo.UploadFile;
 import com.example.konnect_backend.domain.ai.repository.AnalysisRequestLogRepository;
-import com.example.konnect_backend.domain.ai.service.AnalysisHistoryService;
-import com.example.konnect_backend.domain.ai.service.pipeline.module.*;
+import com.example.konnect_backend.domain.ai.service.history.AnalysisHistoryService;
+import com.example.konnect_backend.domain.ai.service.module.*;
 import com.example.konnect_backend.domain.ai.service.prompt.management.PromptLoader;
 import com.example.konnect_backend.domain.ai.service.textextractor.TextExtractorFacade;
 import com.example.konnect_backend.domain.ai.type.TargetLanguage;
 import com.example.konnect_backend.domain.user.entity.Device;
 import com.example.konnect_backend.domain.user.entity.User;
-import com.example.konnect_backend.domain.user.entity.status.Language;
 import com.example.konnect_backend.domain.user.entity.status.UsageType;
 import com.example.konnect_backend.domain.user.repository.DeviceRepository;
 import com.example.konnect_backend.domain.user.repository.UserRepository;
@@ -28,7 +28,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.slf4j.MDC;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
