+++ src/main/java/com/example/konnect_backend/domain/ai/service/prompt/management/PromptManagementService.java
@@ -5,9 +5,9 @@ import com.example.konnect_backend.domain.ai.dto.internal.PromptSummary;
 import com.example.konnect_backend.domain.ai.dto.internal.PromptTemplateWithModelName;
 import com.example.konnect_backend.domain.ai.dto.request.RunPromptRequest;
 import com.example.konnect_backend.domain.ai.dto.response.*;
-import com.example.konnect_backend.domain.ai.entity.AiModel;
-import com.example.konnect_backend.domain.ai.entity.PromptSlot;
-import com.example.konnect_backend.domain.ai.entity.PromptTemplate;
+import com.example.konnect_backend.domain.ai.domain.entity.AiModel;
+import com.example.konnect_backend.domain.ai.domain.entity.PromptSlot;
+import com.example.konnect_backend.domain.ai.domain.entity.PromptTemplate;
 import com.example.konnect_backend.domain.ai.infra.GeminiService;
 import com.example.konnect_backend.domain.ai.repository.AiModelRepository;
 import com.example.konnect_backend.domain.ai.repository.PromptTemplateRepository;
