+++ src/main/java/com/example/konnect_backend/domain/ai/service/textextractor/TextExtractor.java
@@ -1,7 +1,7 @@
 package com.example.konnect_backend.domain.ai.service.textextractor;
 
-import com.example.konnect_backend.domain.ai.model.vo.TextExtractionResult;
-import com.example.konnect_backend.domain.ai.model.vo.UploadFile;
+import com.example.konnect_backend.domain.ai.domain.vo.TextExtractionResult;
+import com.example.konnect_backend.domain.ai.domain.vo.UploadFile;
 import com.example.konnect_backend.domain.ai.type.FileType;
 
 public interface TextExtractor {
